# docs/swagger.yaml
basePath: /
definitions:
...
  handlers.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret
        type: string
    type: object
  handlers.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
    type: object
...
  handlers.RegisterUserRequest:
    properties:
      email:
        example: ivan@example.com
        type: string
      password:
        example: secret123
        type: string
      role:
        example: customer
        type: string
    type: object
...
  models.Computer:
    properties:
      id:
        type: integer
      pc_number:
        type: integer
      status:
        $ref: '#/definitions/models.ComputerStatus'
    type: object
  models.ComputerStatus:
    enum:
    - free
    - busy
    type: string
    x-enum-varnames:
    - Free
    - Busy
...
  models.User:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      role:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample API with Swagger
  title: API Computer-club
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Проверяет email и пароль, возвращает JWT-токен
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный вход
          schema:
            $ref: '#/definitions/handlers.LoginResponse'
        "400":
          description: Некорректный запрос (невалидный JSON)
          schema:
            type: string
        "401":
          description: Ошибка авторизации - неверный email или пароль
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Авторизация пользователя
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя с указанными данными
      parameters:
      - description: Данные для регистрации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно зарегистрирован
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Некорректные данные (пустое имя, email, короткий пароль, неверная
            роль)
          schema:
            type: string
        "409":
          description: Пользователь уже существует
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
  /computers:
    get:
      description: Возвращает список всех компьютеров и их статусы (доступно только
        для администраторов)
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            items:
              $ref: '#/definitions/models.Computer'
            type: array
        "403":
          description: Ошибка доступа - недостаточно прав
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Получение статуса компьютеров
      tags:
      - computers
    post:
      description: Добавляет компьютер в базу данных (доступно только для администраторов)
      produces:
      - application/json
      responses:
        "200":
          description: ': "Компьютер успешно добавлен'
          schema:
            type: string
        "403":
          description: Ошибка доступа - недостаточно прав
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Добавляет компьютер
      tags:
      - computers
  /computers/{id}:
    delete:
      description: Удаляет компьютер из базы данных (доступно только для администраторов)
      produces:
      - application/json
      responses:
        "200":
          description: ': "Компьютер успешно удален'
          schema:
            type: string
        "400":
          description: Некорректный ID компьютера
          schema:
            type: string
        "403":
          description: Ошибка доступа - недостаточно прав
          schema:
            type: string
        "404":
          description: Компьютер не найден
          schema:
            type: string
        "409":
          description: Компьютер занят
          schema:
            type: string
        "500":
          description: Внутренняя ошибка сервера
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Удаляет компьютер
      tags:
      - computers
...
swagger: "2.0"